+++ lib/hexapdf/font/true_type/builder.rb
@@ -59,10 +59,12 @@ module HexaPDF
           tables = tables.sort
 
           tables.each do |tag, data|
+            original_data_length = data.length
             table_checksum = Table.calculate_checksum(data)
+            data << "\0" * (4 - original_data_length % 4) if original_data_length % 4 != 0
             # tag, offset, data.length are all 32bit uint, table_checksum for header and body
             checksum += tag.unpack('N').first + 2 * table_checksum + offset + data.length
-            font_data << [tag, table_checksum, offset, data.length].pack('a4N3')
+            font_data << [tag, table_checksum, offset, original_data_length].pack('a4N3')
             offset += data.length
           end